+++ kie-wb-common-dmn/kie-wb-common-dmn-client/pom.xml
@@ -127,6 +127,11 @@
       <artifactId>kie-wb-common-dynamic-forms-client</artifactId>
     </dependency>
 
+    <dependency>
+      <groupId>org.kie.workbench.widgets</groupId>
+      <artifactId>kie-wb-common-ui</artifactId>
+    </dependency>
+
     <!-- DMN Editor -->
     <dependency>
       <groupId>org.kie.workbench</groupId>
