+++ kie-wb-common-widgets/kie-wb-common-ui/src/main/java/org/kie/workbench/common/widgets/client/cards/frame/CardFrameComponent.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-package org.kie.workbench.common.dmn.client.editors.common.cards.frame;
+package org.kie.workbench.common.widgets.client.cards.frame;
 
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
 
 import elemental2.dom.HTMLElement;
 import org.jboss.errai.ui.client.local.api.elemental2.IsElement;
-import org.kie.workbench.common.dmn.client.editors.common.cards.CardComponent;
+import org.kie.workbench.common.widgets.client.cards.CardComponent;
 import org.uberfire.client.mvp.UberElemental;
 
-import static org.kie.workbench.common.dmn.client.editors.common.cards.CardComponent.DEFAULT_FUNCTION;
+import static org.kie.workbench.common.widgets.client.cards.CardComponent.DEFAULT_FUNCTION;
 
 public class CardFrameComponent {
