+++ kie-wb-common-widgets/kie-wb-common-ui/src/main/java/org/kie/workbench/common/widgets/client/cards/frame/CardFrameComponentView.less
